fixed formula

+++ src/ru/progwards/java1/lessons/basics/AccuracyDoubleFloat.java
@@ -3,12 +3,12 @@ package ru.progwards.java1.lessons.basics;
 public class AccuracyDoubleFloat {
     public static double volumeBallDouble(double radius) {
         final double PI = 3.14;
-        return 4 / 3 * PI * radius * radius * radius;
+        return 4.0 / 3.0 * PI * radius * radius * radius;
     }
 
     public static float volumeBallFloat(float radius) {
         final float PI = 3.14f;
-        return 4 / 3 * PI * radius * radius * radius;
+        return (float)(4.0 / 3.0 * PI * radius * radius * radius);
     }
 
     public static double calculateAccuracy(double radius) {
